Renamed secretWordUtils to wordUtils
